requirements.txt
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
dicttoxml==1.7.4
filetype==1.0.7
future==0.18.2
GeoAlchemy2==0.8.4
geolink-formatter==1.5.2
hupper==1.10.2
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.2
jsonschema==3.2.0
lxml==4.5.1
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.16.6
PasteDeploy==2.1.0
plaster==1.0
plaster-pastedeploy==0.7
psycopg2-binary==2.8.5
pyconizer==0.1.4
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
PyPDF2==1.26.0
pyproj==1.9.6
pyramid==1.10.4
pyramid-debugtoolbar==4.6.1
pyramid-jinja2==2.8
pyramid-mako==1.1.0
pyramid_oereb==1.7.6
pyreproj==1.0.1
pyrsistent==0.16.0
PyYAML==5.3.1
repoze.lru==0.7
requests==2.24.0
Shapely==1.6.4.post1
simplejson==3.17.0
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.6
tqdm==4.47.0
transaction==3.0.0
translationstring==1.4
urllib3==1.25.9
venusian==3.0.0
waitress==1.4.4
WebOb==1.8.6
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==5.1.0
zope.sqlalchemy==1.3
